requirements.txt
Flask==2.0.2
Flask-Cors==3.0.10
Pillow==8.3.1
click==8.0.1
gunicorn==20.1.0
itsdangerous==2.0.1
Jinja2==3.0.1
jinja2-time==0.2.0
MarkupSafe==2.0.1
SQLAlchemy==1.3.17
Werkzeug==2.0.1
pymongo==3.12.0